+++ Operations/Source stage Pipeline
