+++ Payment_(Visa&Paypal)/scansAmexCards.java
